% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/foliate.R
\name{foliate}
\alias{foliate}
\title{Stick things on sticks}
\usage{
foliate(x, n, min.depth = 3, max.depth, ...)
}
\arguments{
\item{x}{A \code{plant} object generated by \code{\link{germinate}}.}

\item{n}{The number of objects to stick to branches.}

\item{min.depth}{The tier of branches below which things will not be stuck to
branches. Default is \code{3}.}

\item{max.depth}{The tier of branches above which things will not be stuck to
branches. If not provided, \code{max.depth} will be set to the outermost
tier of branches.}

\item{...}{Additional arguments passed to \code{points}.}
}
\value{
Nothing is returned. Rather, \code{foliate} is designed to be called
 after a \code{plant} has been plotted, and will add randomly located points
 to the plot with the desired \code{\link{pch}}, colour, size etc.
}
\description{
Spring has sprung. Trees look dead if they don't have leaves, so enfoliate
them. Or, if you prefer, add squirrels and or nuts.
}
\examples{
# Make a seed
s <- seed(50, 10, min.branch.length=0, max.branch.length=5,
         min.trunk.height=5, max.trunk.height=8)

# Germinate it
g <- germinate(s, trunk.width=15, col='peachpuff4')

# Foliate it
leafygreens <- colorRampPalette(paste0('darkolivegreen', c('', 1:4)))(100)
foliate(g, 2000, 4, pch=24:25, col=NA, cex=1.2, bg=paste0(leafygreens, '50'))

# Or add some cherry blossoms
g <- germinate(s, trunk.width=15, col='burlywood4')
floralpinks <- colorRampPalette(c('mistyrose', 'hotpink'))(30)
foliate(g, 3000, 4, pch=24:25, cex=1.5, bg=paste0(floralpinks, '20'), col=NA)
}
\seealso{
\code{\link{germinate}}
}

