% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.xy.R
\name{get.xy}
\alias{get.xy}
\title{Calculate x and y coordinates in a cartesian plane.}
\usage{
get.xy(a, d, x0, y0)
}
\arguments{
\item{a}{Angle from (\code{x0}, \code{y0}).}

\item{d}{Distance from (\code{x0}, \code{y0}).}

\item{x0}{x coordinate from which \code{d} and \code{a} are measured.}

\item{y0}{y coordinate from which \code{d} and \code{a} are measured.}
}
\value{
a \code{data.frame} with coordinates of the desired point.
}
\description{
Calculate x and y coordinates in a cartesian plane, given distance and angle
from a given origin.
}

