% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income}
\alias{income}
\title{Edmonton Income data
A dataset with Edmonton income data.}
\format{
data frame
}
\source{
<https://data.edmonton.ca/Census/2016-Census-Population-by-Household-Income-Neighbo/jkjx-2hix/data_preview>
}
\usage{
income
}
\value{
`tibble`
}
\description{
Edmonton Income data
A dataset with Edmonton income data.
}
\keyword{datasets}
