% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parks}
\alias{parks}
\title{Edmonton Parks data
A dataset with Edmonton parks data.}
\format{
data frame
}
\source{
<https://data.edmonton.ca/Recreation/Parks-in-Edmonton/gdd9-eqv9/data_preview>
}
\usage{
parks
}
\value{
`tibble`
}
\description{
Edmonton Parks data
A dataset with Edmonton parks data.
}
\keyword{datasets}
