% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Edmonton Population data
A dataset with Edmonton population data.}
\format{
data frame
}
\source{
<https://data.edmonton.ca/Census/2016-Census-Population-by-Household-Income-Neighbo/jkjx-2hix/data_preview>
}
\usage{
population
}
\value{
`tibble`
}
\description{
Edmonton Population data
A dataset with Edmonton population data.
}
\keyword{datasets}
