% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{seed}
\alias{seed}
\title{Generate random seeds.}
\usage{
seed(n, max.depth, min.trunk.height = 10, max.trunk.height = 30,
  min.branch.length = 3, max.branch.length = 10)
}
\arguments{
\item{n}{A measure of branchiness. How many terminal branches will this seed 
produce?}

\item{max.depth}{How deep will the hierarchy of branches be? In the wild, 
trees never have more than 25 tiers of branches, nor less than 3, so please 
observe that law of nature.}

\item{min.trunk.height}{How short might the trunk be?}

\item{max.trunk.height}{How tall might the trunk be?}

\item{min.branch.length}{How short might each segment be?}

\item{max.branch.length}{How long might each branch segment be?}
}
\value{
a \code{seed} object containing \code{branches} and \code{lengths}, 
  and suitable for input to \code{\link{germinate}}.
}
\description{
Generate a random seed whose DNA will shape its development upon germination.
}
\examples{
s <- seed(50, 10, min.branch.length=0, max.branch.length=5, 
         min.trunk.height=5, max.trunk.height=8)
         
head(s)
}
\references{
Random sampling of branch tips uses an approach described by 
  Stack Overflow user @BrodieG (Brodie Gaslam) posted 
  \href{http://stackoverflow.com/a/30781090/489704}{here} and available as a 
  Gist \href{https://gist.github.com/brodieG/7fecc0dcbc0cfa53a76a}{here}.
}
\seealso{
\code{\link{germinate}}
}

