% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeg}
\alias{yeg}
\title{yeg tree data}
\format{
data frame
}
\source{
<https://data.edmonton.ca/Environmental-Services/Trees/eecg-fc54/about_data>
}
\usage{
yeg
}
\value{
`tibble`
}
\description{
Edmonton tree data.
}
\keyword{datasets}
