% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeg_neighbourhoods}
\alias{yeg_neighbourhoods}
\title{Yeg Neighborhood data}
\format{
dara frame
A data frame with 7,240 rows and 60 columns:
}
\source{
<https://data.edmonton.ca/Geospatial-Boundaries/Neighbourhoods-and-Wards/gihh-utrc/about_data>
}
\usage{
yeg_neighbourhoods
}
\value{
`sf`
}
\description{
A dataset with Edmonton neighborhood names and polygons coordinates.
}
\keyword{datasets}
