% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/squirrels.R
\name{squirrels}
\alias{squirrels}
\title{Release squirrels.}
\usage{
squirrels(x, branches, pos, left = "L", right = "R", plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{plant} object generated by \code{\link{germinate}}.}

\item{branches}{A binary coded vector of addresses of branches upon which
squirrels (for example) will be placed. The trunk should be coded as
\code{0}, while all other elements of \code{branches} may only contain the
characters given in args \code{left} and \code{right}. E.g. \code{'L'} is
the first branch to the left stemming from the top of the trunk, while
\code{'R'} is the first branch to the right, stemming from the top of the
trunk; \code{'LR'} would be a branch forking to the right from the end of
the first branch that forked left off the trunk.}

\item{pos}{A vector whose elements describe the position of squirrels (for
example), expressed as the distance from the beginning of the branch. This
vector must be the same length as \code{branches}, and its elements
correspond to the elements of \code{branches}.}

\item{left}{The character used to represent left-turning branches in the
\code{branches} vector (default is \code{'L'}).}

\item{right}{The character used to represent right-turning branches in the
\code{branches} vector (default is \code{'R'}). This must not be \code{'L'}.}

\item{plot}{Should the points be added to the current plot?}

\item{...}{Additional arguments controlling the appearance of squirrels (for
example), passed to \code{points}.}
}
\value{
A \code{data.frame} with branch addresses (\code{branches}), squirrel
 positions (\code{pos}), and the corresponding \code{x} and \code{y}
 coordinates. If \code{plot} is \code{TRUE}, the points will be added to the
 current plot.
}
\description{
Add squirrels (for example) at specific locations defined by branch address
and distance along branch.
}
\examples{
# From http://stackoverflow.com/q/28163979/489704
g <- germinate(list(trunk.height=32,
                    branches=c(1, 2, 11, 12, 121, 122),
                    lengths=c(21, 19, 5, 12, 6, 2)),
               left='1', right='2', angle=40)

xy <- squirrels(g, c(0, 1, 121, 1, 11), pos=c(23, 12, 4, 2, 1),
                left='1', right='2', pch=21, bg='white', cex=3, lwd=2)
text(xy$x, xy$y, labels=seq_len(nrow(xy)), font=2)
legend('bottomleft', bty='n',
       legend=paste(seq_len(nrow(xy)),
                    c('FluffyTail', 'Ginger', 'NutCracker', 'SuperSquirrel',
                      'ChipnDale'), sep='. '))
}
\seealso{
\code{\link{germinate}} \code{\link{foliate}}
}

