\name{trees-package}
\alias{trees-package}
\alias{trees}
\docType{package}
\title{
The miracle of binary life.
}
\description{
Farmville has come to R. Create seeds, plant them, watch them grow.
}
\details{
\tabular{ll}{
Package: \tab trees\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-30\cr
License: \tab Copyright Super-plus\cr
}
Take a look at \code{\link{seed}}, \code{\link{germinate}}, and \code{\link{foliate}}.
}
\author{
John Baumgartner

Maintainer: John Baumgartner <johnbaums@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
